refactor: moved service of employee to employee folder
